Disable VTI file check on MacOS
For some reason I do not understand, the VTI files are binary different,
even though their *contents* match. This was confirmed by downloading
the output files as test artifacts and calculating their hashes on my
Ubuntu machine (they match the values from the test logs). Then, I
loaded the binary VTI files in ParaView and stored them as ASCII VTK
files and compared those again, which are identical.

Thus, my suspicion is that there is a difference in the compression
algorithm used between MacOS and Ubuntu/Windows, which causes the binary
files to differ, while the contents are actually identical.